requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
faker==20.1.0



# Vector databases
chromadb==0.4.18
qdrant-client==1.7.0
elasticsearch==8.11.0

# LLM/ML/AI and embeddings
sentence-transformers==2.2.2
huggingface_hub==0.23.4
openai==1.3.5
# torch>=2.1.1
# transformers>=4.35.2

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Text processing
nltk==3.8.1
spacy>=3.6.0

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.0.3
email-validator==2.2.0
requests==2.31.0

# HTTP client
httpx==0.25.2
sse-starlette==1.6.5  # For Server-Sent Events streaming

# Celery and Queue System
celery==5.3.4
redis==5.0.1
kombu==5.3.4
billiard==4.2.0
flower==2.0.1  # Celery monitoring
celery-types==0.18.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
mypy==1.7.1

# Optional: For production deployment
gunicorn==21.2.0

# Logging
loguru==0.7.2

numpy==1.24.0